src/third_party/blink/web_tests/virtual/web-components-v0-disabled/html/dialog/form-method-dialog-expected.txt

CONSOLE ERROR: line 177: Uncaught TypeError: host.createShadowRoot is not a function
Tests form submission with method=dialog

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: host.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

 
